Lection 4

+++ 4_2.c
@@ -2,11 +2,11 @@
 
 main()
 {
-    int i, n, fctr;
-    printf("Введите число: "); scanf("%d", &n);
+    unsigned long long int i, n, fctr;
+    printf("Введите число: "); scanf("%lld", &n);
     for (i = (n - 1); i >=1; i--)
         n *=i;
     system("clear");
-    printf("Значение факториала равно: %d\n", n);
+    printf("Значение факториала равно: %lld\n", n);
     return 0;
 }
